requirements.txt
Django>=2.0
PyNaCl>=1.2.1
